Trait proc_quote::Repeat

pub trait Repeat<T: Iterator>: Sealed<T> { }

Defines the behavior of types that can be interpolated inside repeating patterns (#(...)*).

Which types do Repeat

Iterator<T> consumes the iterator, iterating through every element.

Borrow<[T]> (includes Vec, array, and slice) iterates with the slice::iter method, thus not consuming the original data.

ToTokens, interpolates the variable in every iteration.

Which types do NOT Repeat

IntoIterator, to avoid ambiguity (Ex. “Which behavior would have been used for Vec, which implements both IntoIterator and Borrow<[T]>?”; “Which behavior would have been used for [TokenStream], which implements both IntoIterator and ToTokens?”). To use the iterator, you may call IntoIterator::into_iter explicitly.

Ambiguous types that implement at least two of the Repeat traits. In the very unlikely case this happens, disambiguate the type by wrapping it under some structure that only implements the trait you desire to use.